Fix 1-glibc build after be61b9493d38032519e596f282f9695667402c8d
Which itself fixed:
* [3e2be87832781a29ed67f38f87c1ce3dd4c1b866](https://sourceware.org/git/?p=glibc.git;a=commit;h=3e2be87832781a29ed67f38f87c1ce3dd4c1b866)
* [5dfbc3c43ecc1bcfc760a032c91bb002660051bc](https://sourceware.org/git/?p=glibc.git;a=commit;h=5dfbc3c43ecc1bcfc760a032c91bb002660051bc)

When building, the compiler used for the configure step is different than
the one used for the make step (0-gcc). The latter can't link against
libgcc_s, whereas the first one can, making that configure test, that sets
have-libgcc_s to yes, succeed though it should actually fail. Overwrite
that by deactivating the part that does not work for 0-gcc.